image-api: Fix linting errors after adding `sbt-tpolecat`

+++ image-api/src/main/scala/no/ndla/imageapi/service/search/ImageSearchService.scala
@@ -136,7 +136,7 @@ trait ImageSearchService {
       val sizeFilter = settings.minimumSize match {
         case Some(size) =>
           Some(
-            nestedQuery("imageFiles", rangeQuery("imageFiles.fileSize").gte(size))
+            nestedQuery("imageFiles", rangeQuery("imageFiles.fileSize").gte(size.toLong))
           )
         case _ => None
       }
